Move azure-application-insights
