+++ packages/azure-application-insights/.eslintrc.js
