+++ packages/azure-application-insights/README.md
