+++ packages/azure-application-insights/appInsights.js
