+++ packages/azure-application-insights/db.js
